Navi

Description

This module provides the core application type and logic.

Entry point

runNavi :: forall env m. (HasCallStack, HasEvents env, HasLogEnv env, HasLogQueue env, HasNoteQueue env, MonadAsync m, MonadHandleWriter m, MonadIORef m, MonadLoggerNamespace m, MonadMask m, MonadNotify m, MonadSTM m, MonadSystemInfo m, MonadTerminal m, MonadThread m, MonadReader env m) => m Void

Entry point for the application.

Application Types

newtype NaviT e m a

NaviT is the core type used to run the application.

Constructors

MkNaviT (ReaderT e m a)

runNaviT :: NaviT env m a -> env -> m a

Runs NaviT.