Navi.Config

Description

This modules provides functionality for parsing configuration data from a toml file.

Config

readConfig :: (MonadMutRef ref m, MonadShell m, MonadUnliftIO m) => FilePath -> m (Config ref)

Parses the provided toml file into a Config. Throws ConfigErr if anything goes wrong.

data ConfigErr

ConfigErr represents the errors we can encounter when attempting to parse a config file.

Constructors

FileErr !SomeException

TomlError !TOMLError

NoEvents

data Config ref

Config holds the data from ConfigToml once it has been processed (e.g., all user defined Events are parsed).

Constructors

MkConfig

Fields

events :: !(NonEmpty (AnyEvent ref))

The notification events.

logging :: !Logging

Logging configuration.

noteSystem :: !NoteSystem

The notification system to use.

Logging

data Logging

Logging configuration.

Constructors

MkLogging

Fields

severity :: !(Maybe Severity)

Determines the log level.

location :: !(Maybe LogLoc)

Deterines the log location (i.e. file or stdout).

data LogLoc

Log location configuration.

Constructors

DefPath

Stdout

File !FilePath

Note System

data NoteSystem

Configuration for notification systems.

Constructors

DBus
For use with a running notification server that receives messages via DBus.

NotifySend
For use with the notify-send tool.