Navi.Config.Toml

Description

This module provides functionality for decoding data from a toml configuration file.

Documentation

data ConfigToml

ConfigToml holds the data that is defined in the configuration file.

Constructors

MkConfigToml

Fields

logToml :: !(Maybe Logging)

noteSystemToml :: !(Maybe NoteSystem)

singleToml :: ![SingleToml]

multipleToml :: ![MultipleToml]

batteryPercentageToml :: !(Maybe BatteryPercentageToml)

batteryStatusToml :: !(Maybe BatteryStatusToml)

netInterfacesToml :: ![NetInterfacesToml]