Navi.Config.Types

Description

Provides configuration types.

Config

data Config

Config holds the data from ConfigToml once it has been processed (e.g., all user defined Events are parsed).

Constructors

MkConfig

Fields

events :: NonEmpty AnyEvent

The notification events.

logging :: Logging

Logging configuration.

noteSystem :: NoteSystem 'ConfigPhaseToml

The notification system to use.

data ConfigErr

ConfigErr represents the errors we can encounter when attempting to parse a config file.

Constructors

FileErr SomeException

TomlError TOMLError

NoEvents

Logging

data Logging

Logging configuration.

Constructors

MkLogging

Fields

severity :: Maybe LogLevel

Determines the log level.

location :: Maybe LogLoc

Determines the log location (i.e. file or stdout).

sizeMode :: Maybe FilesSizeMode

Determines whether to warn/delete large log files.

data LogLoc

Log location configuration.

Constructors

DefPath

Stdout

File OsPath

data FilesSizeMode

Determines what to do if the log file surpasses the given size threshold.

Constructors

FilesSizeModeWarn (Bytes 'B Natural)
Print a warning.

FilesSizeModeDelete (Bytes 'B Natural)
Delete the file.

defaultLogging :: Logging

Default logging i.e. log errors and use the default path.

defaultSizeMode :: FilesSizeMode

Since: 0.1

Note System

data NoteSystem (p :: ConfigPhase)

Configuration for notification systems.

Constructors

AppleScript
For use with osx.

DBus (DBusF p)
For use with a running notification server that receives messages via DBus.

NotifySend
For use with the notify-send tool.

defaultNoteSystem :: NoteSystem 'ConfigPhaseToml

Default notification system i.e. DBus for linux, AppleScript for osx.