Navi.Env.Core

Description

This module provides the core Env type for Navi.

HasX-style Typeclasses

class HasEvents ref env | env -> ref where

Retrieves the events.

Methods

getEvents :: env -> NonEmpty (AnyEvent ref)

class HasLogEnv env where

Retrieves the log environment.

Methods

getLogEnv :: env -> LogEnv

setLogEnv :: LogEnv -> env -> env

overLogEnv :: (LogEnv -> LogEnv) -> env -> env

class HasLogContexts env where

Retrieves the log context.

Methods

getLogContexts :: env -> LogContexts

setLogContexts :: LogContexts -> env -> env

overLogContexts :: (LogContexts -> LogContexts) -> env -> env

class HasLogNamespace env where

Retrieves the log namespace.

Methods

getLogNamespace :: env -> Namespace

setLogNamespace :: Namespace -> env -> env

overLogNamespace :: (Namespace -> Namespace) -> env -> env

class HasLogQueue env where

Retrieves the log queue.

Methods

getLogQueue :: env -> NaviQueue (NaviLog, Namespace)

class HasNoteQueue env where

Retrieves the note queue.

Methods

getNoteQueue :: env -> NaviQueue NaviNote

Concrete Env

data Env ref

Env holds all of our environment data that is used while running navi.

Constructors

MkEnv

Fields

events :: !(NonEmpty (AnyEvent ref))

logEnv :: !LogEnv

logCtx :: !LogContexts

logNamespace :: !Namespace

logQueue :: !(NaviQueue (NaviLog, Namespace))

noteQueue :: !(NaviQueue NaviNote)