Navi.Event

Description

This module provides functionality for handling events.

Event type

data Event ref result

Event represents sending notifications. An event will:

Query for information (i.e. run a shell command).

Parse the result.

Raise an alert if the result matches some condition.

Constructors

MkEvent

Fields

name :: !Text

The name of this event.

serviceType :: !(ServiceType result)

The service to run.

pollInterval :: !PollInterval

How often to poll for this event, in seconds.

raiseAlert :: result -> Maybe NaviNote

Conditionally raises an alert based on the result.

repeatEvent :: !(RepeatEvent ref result)

Determines how we handle repeat alerts.

errorNote :: !(ErrorNote ref)

Determines how we handle errors.

data AnyEvent ref where

Existentially quantifies result type on an Event. Used so that we can store different events in the same list.

Constructors

MkAnyEvent :: (Eq result, Show result) => Event ref result -> AnyEvent ref

runEvent :: (HasLogNamespace env, HasLogQueue env, MonadLogger m, MonadQueue m, MonadReader env m, MonadSystemInfo m, Show result) => Event ref result -> m result

Runs an event, i.e.,

Queries the system via MonadSystemInfo.

Returns the parsed result.

Results

data EventError

Represents an error when querying an Event.

Constructors

MkEventError

Fields

name :: !Text

The name of the event.

short :: !Text

Short description of the error.

long :: !Text

Long description of the error.

Caching previous events/errors

data RepeatEvent ref a

Determines if we are allowed to send off duplicate notifications simultaneously. If we are not, then NoRepeats holds the last trigger so that we can detect duplicates.

Constructors

NoRepeats !(ref (Maybe a))

AllowRepeats

data ErrorNote ref

Determines if we should send notifications for errors and, if so, if we allow repeats.

Constructors

NoErrNote

AllowErrNote !(RepeatEvent ref ())

blockRepeat :: (Eq a, HasLogQueue env, HasLogNamespace env, MonadLogger m, MonadMutRef ref m, MonadQueue m, MonadReader env m, Show a) => RepeatEvent ref a -> a -> m Bool

Determines if we should block the event. The semantics are:

AllowRepeats: never block (returns False).

NoRepeats: block only if the parameter a equals the previous a stored in our ref.

blockErr :: (HasLogNamespace env, HasLogQueue env, MonadLogger m, MonadMutRef ref m, MonadQueue m, MonadReader env m) => ErrorNote ref -> m Bool

Determines if we should block the error event. The semantics are:

NoErrNote: always block (returns True).

AllowErrNote AllowRepeats: never block (returns False).

AllowErrNote NoRepeats: block only if we have sent a notifcation for this error before.

updatePrevTrigger :: (Eq a, MonadMutRef ref m) => RepeatEvent ref a -> a -> m ()

If the reference is NoRepeats then we overwrite the previous reference with the new parameter. Otherwise we do nothing.