Navi.Event.Types

Description

This module provides types for defining notification events.

Documentation

data Event ref result

Event represents sending notifications. An event will:

Query for information (i.e. run a shell command).

Parse the result.

Raise an alert if the result matches some condition.

Constructors

MkEvent

Fields

name :: !Text

The name of this event.

serviceType :: !(ServiceType result)

The service to run.

pollInterval :: !PollInterval

How often to poll for this event, in seconds.

raiseAlert :: result -> Maybe NaviNote

Conditionally raises an alert based on the result.

repeatEvent :: !(RepeatEvent ref result)

Determines how we handle repeat alerts.

errorNote :: !(ErrorNote ref)

Determines how we handle errors.

data AnyEvent ref where

Existentially quantifies result type on an Event. Used so that we can store different events in the same list.

Constructors

MkAnyEvent :: (Eq result, Show result) => Event ref result -> AnyEvent ref

data RepeatEvent ref a

Determines if we are allowed to send off duplicate notifications simultaneously. If we are not, then NoRepeats holds the last trigger so that we can detect duplicates.

Constructors

NoRepeats !(ref (Maybe a))

AllowRepeats

_NoRepeats :: forall ref a ref a. Prism (RepeatEvent ref a) (RepeatEvent ref a) (ref (Maybe a)) (ref (Maybe a))

_AllowRepeats :: forall ref a. Prism' (RepeatEvent ref a) ()

data ErrorNote ref

Determines if we should send notifications for errors and, if so, if we allow repeats.

Constructors

NoErrNote

AllowErrNote !(RepeatEvent ref ())

_NoErrNote :: forall ref. Prism' (ErrorNote ref) ()

_AllowErrNote :: forall ref ref. Prism (ErrorNote ref) (ErrorNote ref) (RepeatEvent ref ()) (RepeatEvent ref ())

data EventError

Represents an error when querying an Event.

Constructors

MkEventError

Fields

name :: !Text

The name of the event.

short :: !Text

Short description of the error.

long :: !Text

Long description of the error.