Navi.Services.Battery.Percentage.Toml

Description

This module provides toml configuration for the battery percentage service.

Documentation

data BatteryPercentageToml

TOML for the battery percentage service.

Constructors

MkBatteryPercentageToml

Fields

alerts :: NonEmpty BatteryPercentageNoteToml

All alerts for this service.

pollInterval :: Maybe PollInterval

The poll interval.

repeatEvent :: Maybe RepeatEventToml

Determines how we treat repeat alerts.

errorNote :: Maybe ErrorNoteToml

Determines how we handle errors.

app :: RunApp BatteryApp

Determines how we should query the system for battery information.

data BatteryPercentageNoteToml

TOML for each individual battery percentage.

Constructors

MkBatteryPercentageNoteToml

Fields

percentage :: Percentage

The percentage for this alert.

urgency :: Maybe UrgencyLevel

The urgency for this alert.

mTimeout :: Maybe Timeout

The timeout for this alert.