Navi.Services.Battery.Status.Toml

Description

This module provides toml configuration for the battery status service.

Documentation

data BatteryStatusToml

TOML for the battery status service.

Constructors

MkBatteryStatusToml

Fields

app :: BatteryApp

Determines how we should query the system for battery information.

pollInterval :: Maybe PollInterval

The poll interval.

repeatEvent :: Maybe RepeatEventToml

Determines how we treat repeat alerts.

errorNote :: Maybe ErrorNoteToml

Determines how we handle errors.

mTimeout :: Maybe Timeout

The timeout for this alert.