Navi.Services.Network.NetInterfaces.Toml

Description

This module provides toml configuration for the network connectivity service.

Documentation

data NetInterfacesToml

TOML for the network connectivity service.

Constructors

MkNetInterfacesToml

Fields

app :: NetInterfaceApp

Determines how we should query the system for network information.

deviceName :: Text

The name of the network device. For "standard" formats like ifconfig or NetworkManager, this might be something like wlp0s20f3 or enp0s31f6.

pollInterval :: Maybe PollInterval

The poll interval.

repeatEvent :: Maybe RepeatEventToml

Determines how we treat repeat alerts.

errorNote :: Maybe ErrorNoteToml

Determines how we handle errors.

mTimeout :: Maybe Timeout

The timeout for this alert.