Navi.Services.Types

Description

This module provides the ServiceType type.

Documentation

data ServiceType result where

ServiceType describes all implemented services. It provides several built-in services for querying and parsing system information (from the pythia package), and then Single and Custom for custom services. It is a GADT so we can link each service with its result type.

Constructors

BatteryPercentage :: BatteryApp -> ServiceType Battery

BatteryStatus :: BatteryApp -> ServiceType BatteryStatus

NetworkInterface :: Device -> NetInterfaceApp -> ServiceType NetInterface

Custom :: Command -> CommandResultParser -> ServiceType CommandResult

_BatteryPercentage :: result ~ Battery => Prism' (ServiceType result) BatteryApp

_BatteryStatus :: result ~ BatteryStatus => Prism' (ServiceType result) BatteryApp

_Custom :: result ~ CommandResult => Prism' (ServiceType result) (Command, CommandResultParser)

_NetworkInterface :: result ~ NetInterface => Prism' (ServiceType result) (Device, NetInterfaceApp)