Navi.Utils

Description

Provides utilities.

Since: 0.1

TOML

Decoding utils

getFieldOptArrayOf :: DecodeTOML a => Text -> Decoder [a]

Decodes an optional list. This is morally

getFieldOptWith (getArrayOf tomlDecoder) :: Text -> Decoder (Maybe [a])

except we return an empty list when the key is missing rather than Nothing.

Since: 0.1

Specific decoders

commandDecoder :: Decoder Command

TOML decoder for Command with field name "command".

Since: 0.1

runAppDecoder :: Decoder a -> Decoder (RunApp a)

TOML decoder for optional RunApp with field name "app". If the field is not found then we decode Many.

Since: 0.1

urgencyLevelOptDecoder :: Decoder (Maybe UrgencyLevel)

TOML decoder for optional UrgencyLevel with field name "urgency".

Since: 0.1