Navi.Config.Toml

Description

This module provides functionality for decoding data from a toml configuration file.

Documentation

data ConfigToml

ConfigToml holds the data that is defined in the configuration file.

Constructors

MkConfigToml

Fields

batteryPercentageToml :: Maybe BatteryPercentageToml

batteryStatusToml :: Maybe BatteryStatusToml

customToml :: [CustomToml]

logToml :: Maybe Logging

netInterfacesToml :: [NetInterfacesToml]

noteSystemToml :: Maybe (NoteSystem 'ConfigPhaseToml)