Navi.Data.PollInterval

Description

Provides the PollInterval type.

Since: 0.1

Documentation

newtype PollInterval

Represents how often to poll for service changes, in seconds.

Since: 0.1

Constructors

MkPollInterval

Fields

unPollInterval :: Natural

parsePollInterval :: MonadFail m => Text -> m PollInterval

pollIntervalOptDecoder :: Decoder (Maybe PollInterval)

TOML decoder for optional PollInterval with field name 'poll-interval'.

Since: 0.1

toSleepTime :: PollInterval -> Int

Converts a PollInterval into an Int suitable to be used with threadDelay.

Since: 0.1