Navi.Env.Core

Description

This module provides the core Env type for Navi.

HasX-style Typeclasses

class HasEvents env where

Retrieves the events.

Methods

getEvents :: env -> NonEmpty AnyEvent

class HasLogEnv env where

Retrieves the log environment.

Methods

getLogEnv :: env -> Maybe LogEnv

class HasNoteQueue env where

Retrieves the note queue.

Methods

getNoteQueue :: env -> TBQueue NaviNote

sendNoteQueue :: (HasCallStack, HasNoteQueue env, MonadReader env m, MonadSTM m) => NaviNote -> m ()

Convenience function for retrieving a TBQueue NaviNote from the env and sending the note.

Deriving

newtype TopField a

Used for deriving instances from the top level field name e.g. 'events :: NonEmpty AnyEvent'.

Constructors

MkTopField a

newtype CoreEnvField a

Used for deriving instances for types with a field 'coreEnv :: Env'.

Constructors

MkCoreEnvField a

Concrete Env

data Env

Env holds all of our environment data that is used while running navi.

Constructors

MkEnv

Fields

events :: NonEmpty AnyEvent

logEnv :: Maybe LogEnv

noteQueue :: TBQueue NaviNote

notifySystem :: NoteSystem 'ConfigPhaseEnv