Navi.Env.Core

Description

This module provides the core Env type for Navi.

HasX-style Typeclasses

class HasEvents env where

Retrieves the events.

Methods

getEvents :: env -> NonEmpty AnyEvent

class HasLogEnv env where

Retrieves the log environment.

Methods

getLogEnv :: env -> LogEnv

localLogEnv :: (LogEnv -> LogEnv) -> env -> env

class HasLogQueue env where

Retrieves the log queue.

Methods

getLogQueue :: env -> TBQueue LogStr

class HasNoteQueue env where

Retrieves the note queue.

Methods

getNoteQueue :: env -> TBQueue NaviNote

Concrete Env

data Env

Env holds all of our environment data that is used while running navi.

Constructors

MkEnv

Fields

events :: !(NonEmpty AnyEvent)

logEnv :: !LogEnv

logQueue :: !(TBQueue LogStr)

noteQueue :: !(TBQueue NaviNote)