Navi.Env.NotifySend

Description

Provides environment for usage with NotifySend.

Documentation

mkNotifySendEnv :: MonadSTM m => Maybe LogEnv -> Config -> m Env

Creates a NotifySendEnv from the provided log types and configuration data.

naviToNotifySend :: NaviNote -> Text

Turns a NaviNote into a string to be sent with the notify-send tool.