Navi.Env.NotifySend

Description

Provides environment for usage with NotifySend.

Documentation

newtype NotifySendEnv

Concrete notify-send environment. Adds the dbus client.

Constructors

MkNotifySendEnv

Fields

coreEnv :: Env

mkNotifySendEnv :: MonadSTM m => LogEnv -> Config -> m NotifySendEnv

Creates a NotifySendEnv from the provided log types and configuration data.

naviToNotifySend :: NaviNote -> Text

Turns a NaviNote into a string to be sent with the notify-send tool.