Navi.Event

Description

This module provides functionality for handling events.

Event type

data Event result trigger

Event represents sending notifications. An event will:

Query for information (i.e. run a shell command).

Parse the result.

Map the result to a trigger.

Raise an alert if the trigger matches some condition.

For most services, the (result -> trigger) map will be trivial i.e. result and trigger will be identical.

For example, custom servicess is Text, and NetInterfaces is NetInterface. In these cases, there is an exact correspondonce between the service query result and the trigger.

But for e.g. battery percentage, the result is Battery (percentage and status) whereas trigger is PercentageData.

See NOTE: [Battery Percentage Result/Trigger] for an explanation on why this might be desirable.

Constructors

MkEvent

Fields

errorNote :: ErrorNote

Determines how we handle errors.

name :: Text

The name of this event.

pollInterval :: PollInterval

How often to poll for this event, in seconds.

raiseAlert :: result -> Maybe (trigger, NaviNote)

Conditionally raises an alert based on the (result -> trigger) mapping.

repeatEvent :: RepeatEvent trigger

Determines how we handle repeat alerts.

serviceType :: ServiceType result

The service to run.

data AnyEvent where

Existentially quantifies result type on an Event. Used so that we can store different events in the same list.

Constructors

MkAnyEvent :: forall trigger result. (Ord trigger, Show result, Show trigger) => Event result trigger -> AnyEvent

runEvent :: (HasCallStack, MonadLoggerNS m env k, MonadSystemInfo m, Show result) => Event result trigger -> m (EventSuccess result trigger)

Runs an event, i.e.,

Queries the system via MonadSystemInfo.

Returns the parsed result.

Results

data EventSuccess result trigger

Holds the Event data used after an event is successfully run.

Since: 0.1

Constructors

MkEventSuccess

Fields

pollInterval :: Maybe PollInterval

result :: result

repeatEvent :: RepeatEvent trigger

raiseAlert :: result -> Maybe (trigger, NaviNote)

data EventError

Represents an error when querying an Event.

Constructors

MkEventError

Fields

name :: Text

The name of the event.

short :: Text

Short description of the error.

long :: Text

Long description of the error.

Caching previous events/errors

data RepeatEvent a

Determines if we are allowed to send off duplicate notifications simultaneously. If we are not, then NoRepeats holds the last trigger so that we can detect duplicates.

Constructors

NoRepeats (IORef (Maybe a))

SomeRepeats (Set a) (IORef (Maybe a))

AllowRepeats

data ErrorNote

Determines if we should send notifications for errors and, if so, if we allow repeats.

Constructors

NoErrNote

AllowErrNote (RepeatEvent ())

blockRepeat :: (Ord a, MonadLoggerNS m env k, MonadIORef m, Show a) => RepeatEvent a -> a -> m Bool

Determines if we should block the event. The semantics are:

AllowRepeats: never block (returns False).

NoRepeats: block only if the parameter a equals the previous a stored in our ref.

blockErr :: (MonadLoggerNS m env k, MonadIORef m) => ErrorNote -> m Bool

Determines if we should block the error event. The semantics are:

NoErrNote: always block (returns True).

AllowErrNote AllowRepeats: never block (returns False).

AllowErrNote NoRepeats: block only if we have sent a notifcation for this error before.

updatePrevTrigger :: (Eq a, MonadIORef m) => RepeatEvent a -> Maybe a -> m ()

If the reference is NoRepeats or SomeRepeats then we overwrite the previous reference with the new parameter. Otherwise we do nothing.