Navi.Services.Battery.Percentage.Toml

Description

This module provides toml configuration for the battery percentage service.

Documentation

data BatteryPercentageToml

TOML for the battery percentage service.

Constructors

MkBatteryPercentageToml

Fields

alerts :: NonEmpty BatteryPercentageNoteToml

All alerts for this service.

app :: BatteryApp

Determines how we should query the system for battery information.

errorNote :: Maybe ErrorNoteToml

Determines how we handle errors.

pollInterval :: Maybe PollInterval

The poll interval.

repeatEvent :: Maybe (MultiRepeatEventToml Percentage)

Determines how we treat repeat alerts. We are given exact percentages, so if the user wants a repeatEvent to apply to a range, they should give the _lower_ bound.

data BatteryPercentageNoteToml

TOML for each individual battery percentage.

Constructors

MkBatteryPercentageNoteToml

Fields

percentage :: PercentageData

The percentage (range) for this alert.

urgency :: Maybe UrgencyLevel

The urgency for this alert.

mTimeout :: Maybe Timeout

The timeout for this alert.

data PercentageData

Exact percentage or range. Equality/Ord is based on the _lower_ part of the range only. This allows us to have a total order and make lookups sensible.

Constructors

PercentageExact Percentage
Exact percentage.

PercentageRange Percentage Percentage
Percentage range. The LHS should be <= RHS.